move TestUtils from tests to main + rename to LocationUtils I wanted to construct a location from a given lat, long and this class appeared to be useful outside of the tests
